+++ src/dsp_tools/commands/xmlupload/stash/stash_circular_references.py
@@ -7,7 +7,7 @@ from lxml import etree
 
 from dsp_tools.commands.xmlupload.models.formatted_text_value import FormattedTextValue
 from dsp_tools.commands.xmlupload.models.permission import Permissions
-from dsp_tools.commands.xmlupload.models.xmlproperty import XMLProperty
+from dsp_tools.commands.xmlupload.models.Value_deserialise import XMLProperty
 from dsp_tools.commands.xmlupload.models.xmlresource import XMLResource
 from dsp_tools.commands.xmlupload.stash.construct_and_analyze_graph import create_info_from_xml_for_graph
 from dsp_tools.commands.xmlupload.stash.construct_and_analyze_graph import generate_upload_order
