+++ test/unittests/commands/xmlupload/test_Value_deserialise.py
@@ -4,7 +4,7 @@ import pytest
 from lxml import etree
 
 from dsp_tools.commands.xmlupload.models.formatted_text_value import FormattedTextValue
-from dsp_tools.commands.xmlupload.models.xmlvalue import XMLValue
+from dsp_tools.commands.xmlupload.models.Value_deserialise import XMLValue
 
 # ruff: noqa: PT009 (pytest-unittest-assertion) (remove this line when pytest is used instead of unittest)
